src/requirements.txt
flask==1.0.2
psycopg2==2.7.7
psycopg2-binary==2.7.7
PyMySQL==0.9.3
SQLAlchemy==1.2.17
requests
google-api-python-client
oauthlib==3.0.1
google-auth
google-auth-oauthlib
google-auth-httplib2
